added background to myprojects page
